\name{fastliclust-package}
\alias{fastliclust-package}
\alias{fastliclust}
\docType{package}
\title{
\packageTitle{fastliclust}
}
\description{
\packageDescription{fastliclust}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fastliclust}
\packageIndices{fastliclust}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fastliclust}

Maintainer: \packageMaintainer{fastliclust}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
