% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastLiclust}
\alias{fastLiclust}
\title{Hierarchical clustering on a linkage matrix}
\usage{
fastLiclust(linkmat, sim, weights, disconnect = 1)
}
\arguments{
\item{linkmat}{A nx2 integer matrix of vertices which are connected.}

\item{sim}{A length n numeric vector of the distances between the connected edges.}

\item{weights}{A n integer vector initialized with all 1, which will contain the number of vertices
summarized under a cluster (necessary in the algorithm.)}
}
\value{
Nothing - the operation is performed in place on the matrix! Note this is highly unstandard
R behavior. To get the results in a useful way, run first \link{\code{crop}} then \link{\code{toHclust}}.
}
\description{
Performs hierarchical clustering (currently only average linkage is supported) on a dataset which is 
only partially interconnected (i.e. between some vertices there is no edge.) This is in particular a
memory-saving alternative on large sparse graphs which are only encoded in the linkage matrix, i.e. have
no nxn distance matrix. Other use cases are sparsely populated matrices which can be extracted directly from 
\code{ff} objects on disk, so no nxn matrix has to be built in RAM.
}
