% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ff_which_matrix}
\alias{ff_which_matrix}
\title{which in ff matrices}
\usage{
ff_which_matrix(X, cond, by = 100)
}
\arguments{
\item{X}{The \code{ff_matrix} to scan}

\item{cond}{The condition to evaluate}

\item{by}{The chunk size.}
}
\value{
X
}
\description{
Returns the indices of matching elements.
}
\examples{
 
mymat <- ff(vmode="double", dim=c(50,50))
mymat[4,5] <- 3
mymat[5,5] <- 4
mymat[46, 46] <- 1

ff_which_matrix(mymat, mymat > 2)

r <- 1.5
ff_which_matrix(mymat, mymat > r)

}

