% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_ff}
\alias{in_ff}
\title{Perform chunked expression within an ff_matrix object}
\usage{
in_ff(X, expr, by = 500, ...)
}
\arguments{
\item{X}{An \code{ff_matrix}}

\item{expr}{The expression to evaluate}

\item{by}{Chunk size}

\item{...}{Further arguments}
}
\value{
X
}
\description{
Perform chunked expression within an ff_matrix object
}
\examples{

mymat <- ff(vmode="double", dim=c(50,50))
mymat[4,5] <- 3
mymat[5,5] <- 4
mymat[46, 46] <- 1
in_ff(mymat, mymat[mymat > 2] <- 7)

mymat

fun <- function(x) 2*x
in_ff(mymat, mymat[mymat > 2] <- fun(mymat[mymat > 2]))

mymat

r<- 8
in_ff(mymat, mymat[mymat > 2] <- r)

mymat

}
