% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastliclust.R
\name{toHclust}
\alias{toHclust}
\title{Convert fastliclust output to hclust-compatible output}
\usage{
toHclust(linkmat, sim)
}
\arguments{
\item{linkmat}{The linkage matrix after processing fastliclust}

\item{sim}{The distance list after processing fastliclust}
}
\value{
a hclust-style object. See hclust for details
}
\description{
Convert fastliclust output to hclust-compatible output
}
