% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastliclust.R
\name{toLinkmat}
\alias{toLinkmat}
\title{Convert distance matrix to input suitable for fastLiclust}
\usage{
toLinkmat(mat, disconnect = 1)
}
\arguments{
\item{mat}{The input distance matrix, as obtained from as.matrix() on a dist object}

\item{disconnect}{The distance value for disconnected features, e.g. 1 in dissimilarity matrices, or 
NA when that value was set to NA before. These points are not represented in the linkage matrix.}
}
\value{
A list with components linkmat (a matrix of connected vertex pairs), 
   sim (the distances associated with the edges), and weights (a counter which denotes the number
   of original vertices summed up in a cluster, initialized to 1.)
}
\description{
Mainly intended to show equivalence with hclust, when one generates a distance
matrix to use with hclust and then with fastLiclust to generate the same dendrogram.
}
